feat(ui): refactor filters with slide and popup variants
