+++ ui/dashboard/src/elements/filter/index.tsx
@@ -1,15 +1,14 @@
-import { ReactNode } from 'react';
-import {
-  IconLaunchOutlined,
-  IconFilterListOutlined
-} from 'react-icons-material-design';
-import { Link } from 'react-router';
-import { useScreen } from 'hooks';
-import { useTranslation } from 'i18n';
-import { cn } from 'utils/style';
 import Button from 'components/button';
 import Icon from 'components/icon';
 import SearchInput from 'components/search-input';
+import { useTranslation } from 'i18n';
+import { ReactNode } from 'react';
+import {
+  IconFilterListOutlined,
+  IconLaunchOutlined
+} from 'react-icons-material-design';
+import { Link } from 'react-router';
+import { cn } from 'utils/style';
 
 interface FilterProps {
   action?: ReactNode;
@@ -17,6 +16,7 @@ interface FilterProps {
   filterCount?: number;
   isShowDocumentation?: boolean;
   className?: string;
+  actionClassName?: string;
   link?: string;
   placeholder?: string;
   name?: string;
@@ -31,6 +31,7 @@ const Filter = ({
   filterCount,
   isShowDocumentation = true,
   className,
+  actionClassName,
   link = '',
   placeholder,
   name,
@@ -38,17 +39,16 @@ const Filter = ({
   onOpenFilter
 }: FilterProps) => {
   const { t } = useTranslation(['common', 'form']);
-  const { lessThanXLScreen } = useScreen();
   return (
     <div
       className={cn(
-        'flex w-full lg:items-center justify-between flex-col lg:flex-row pl-6 pr-6 gap-x-6',
-        { '!flex-row !justify-end': !onSearchChange },
+        'flex flex-wrap w-full lg:items-center justify-between lg:flex-row px-3 sm:px-6 gap-x-6',
+        { '!flex-row lg:!justify-end': !onSearchChange },
         className
       )}
     >
       {onSearchChange && (
-        <div className="w-full max-w-[365px]">
+        <div className="w-full max-w-full lg:max-w-[365px]">
           <SearchInput
             name={name}
             placeholder={placeholder || `${t('form:placeholder-search-input')}`}
@@ -59,14 +59,13 @@ const Filter = ({
       )}
       <div
         className={cn(
-          'flex flex-1 w-full items-center justify-end gap-4 mt-3 lg:mt-0',
-          {
-            'flex-wrap': lessThanXLScreen
-          }
+          'flex flex-wrap flex-1 w-full items-center  justify-start md:justify-end gap-4 mt-3 lg:mt-0',
+          actionClassName
         )}
       >
         {isShowDocumentation && (
           <Link
+            className="hidden sm:block"
             target="_blank"
             to={link}
             onClick={e => {
@@ -83,7 +82,7 @@ const Filter = ({
           <Button
             variant="secondary"
             onClick={onOpenFilter}
-            className="text-gray-700 shadow-border-gray-400 flex-1 lg:flex-none"
+            className="text-gray-700 shadow-border-gray-400 flex-1 px-2 sm:px-4 lg:flex-none"
           >
             <Icon icon={IconFilterListOutlined} size="sm" />
             {t('filter')}
